Extract test helper for checker db functions
I think this might be somewhat idiomatic?
Just need to share some test functions as we're about to test another check.
